fastapi/requirements.txt
fastapi==0.115.5
uvicorn
sqlalchemy
psycopg2-binary
pdfplumber
pandas